+++ CHANGES.rst
@@ -5,8 +5,9 @@ Here you can see the full list of changes between each PostgreSQL-Audit release.
 
 
 0.1.1 (2015-02-17)
-^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^^
 
+- Added __repr__ for activity classes
 - Removed session user name column from activity table
 - Removed application name column from activity table